+++ sdk-media-webrtc/src/main/kotlin/com/pexip/sdk/media/webrtc/WebRtcMediaConnectionFactory.kt
@@ -77,26 +77,6 @@ public class WebRtcMediaConnectionFactory private constructor(
     CameraVideoTrackFactory,
     MediaProjectionVideoTrackFactory {
 
-    @Deprecated(
-        message = "Use WebRtcMediaConnectionFactory.Builder instead.",
-        level = DeprecationLevel.ERROR,
-    )
-    @JvmOverloads
-    public constructor(
-        context: Context,
-        eglBase: EglBase,
-        cameraEnumerator: CameraEnumerator = CameraEnumerator(context),
-        videoDecoderFactory: VideoDecoderFactory = DefaultVideoDecoderFactory(eglBase),
-        videoEncoderFactory: VideoEncoderFactory = DefaultVideoEncoderFactory(eglBase),
-    ) : this(
-        applicationContext = context.applicationContext,
-        eglBase = eglBase,
-        cameraEnumerator = cameraEnumerator,
-        audioDeviceModule = JavaAudioDeviceModule(context),
-        videoDecoderFactory = videoDecoderFactory,
-        videoEncoderFactory = videoEncoderFactory,
-    )
-
     private val factory = PeerConnectionFactory.builder()
         .setAudioDeviceModule(audioDeviceModule)
         .setVideoDecoderFactory(videoDecoderFactory)
@@ -136,22 +116,6 @@ public class WebRtcMediaConnectionFactory private constructor(
             scope = CoroutineScope(),
             signalingDispatcher = signalingDispatcher,
         )
-    }
-
-    @Deprecated(
-        message = "Use createCameraVideoTrack() that also accepts a Callback.",
-        level = DeprecationLevel.ERROR,
-    )
-    override fun createCameraVideoTrack(): CameraVideoTrack {
-        throw UnsupportedOperationException()
-    }
-
-    @Deprecated(
-        message = "Use createCameraVideoTrack() that also accepts a Callback.",
-        level = DeprecationLevel.ERROR,
-    )
-    override fun createCameraVideoTrack(deviceName: String): CameraVideoTrack {
-        throw UnsupportedOperationException()
     }
 
     override fun createCameraVideoTrack(callback: CameraVideoTrack.Callback): CameraVideoTrack {