+++ sdk-media/src/main/kotlin/com/pexip/sdk/media/CameraVideoTrackFactory.kt
@@ -53,38 +53,6 @@ public interface CameraVideoTrackFactory {
      * 2. First back-facing camera
      * 3. First available camera
      *
-     * @return a [CameraVideoTrack]
-     * @throws IllegalStateException if no camera is available
-     * @throws IllegalStateException if [MediaConnectionFactory] has been disposed
-     */
-    @Deprecated(
-        message = "Use createCameraVideoTrack() that also accepts a Callback.",
-        level = DeprecationLevel.ERROR,
-    )
-    public fun createCameraVideoTrack(): CameraVideoTrack
-
-    /**
-     * Creates a [CameraVideoTrack] for specific [deviceName].
-     *
-     * @param deviceName a device name that should be opened
-     * @return a [CameraVideoTrack]
-     * @throws IllegalStateException if [deviceName] is not available
-     * @throws IllegalStateException if [MediaConnectionFactory] has been disposed
-     */
-    @Deprecated(
-        message = "Use createCameraVideoTrack() that also accepts a Callback.",
-        level = DeprecationLevel.ERROR,
-    )
-    public fun createCameraVideoTrack(deviceName: String): CameraVideoTrack
-
-    /**
-     * Creates a [CameraVideoTrack] for the best available camera.
-     *
-     * Best available camera is determined by the following order:
-     * 1. First front-facing camera
-     * 2. First back-facing camera
-     * 3. First available camera
-     *
      * @param callback a callback used to signal various camera events
      * @return a [CameraVideoTrack]
      * @throws IllegalStateException if no camera is available